+++ src/daft-sql/src/modules/config.rs
@@ -1,16 +1,12 @@
 use std::collections::HashMap;
 
-use common_io_config::{HTTPConfig, S3Config};
+use common_io_config::{HTTPConfig, IOConfig, S3Config};
 use daft_core::prelude::{DataType, Field};
-use daft_dsl::{literal_value, null_lit, Expr, ExprRef, LiteralValue};
-use serde::{
-    de::{DeserializeOwned, DeserializeSeed, IntoDeserializer, MapAccess, Visitor},
-    Deserialize, Deserializer,
-};
-use sqlparser::ast::{FunctionArg, FunctionArgOperator};
+use daft_dsl::{literal_value, Expr, ExprRef, LiteralValue};
 
 use super::SQLModule;
 use crate::{
+    error::{PlannerError, SQLPlannerResult},
     functions::{SQLFunction, SQLFunctionArguments, SQLFunctions},
     unsupported_sql_err,
 };
@@ -21,6 +17,8 @@ impl SQLModule for SQLModuleConfig {
     fn register(parent: &mut SQLFunctions) {
         parent.add_fn("S3Config", S3ConfigFunction);
         parent.add_fn("HTTPConfig", HTTPConfigFunction);
+        // parent.add_fn("AzureConfig", AzureConfigFunction);
+        // parent.add_fn("GCSConfig", GCSConfigFunction);
     }
 }
 
@@ -33,15 +31,14 @@ macro_rules! item {
         )
     };
 }
+
 impl SQLFunction for S3ConfigFunction {
     fn to_expr(
         &self,
         inputs: &[sqlparser::ast::FunctionArg],
         planner: &crate::planner::SQLPlanner,
     ) -> crate::error::SQLPlannerResult<daft_dsl::ExprRef> {
-        let deserializer = FunctionArgDeserializer::new(inputs, planner);
-        let cfg = deserializer.deserialize::<S3Config>();
-        println!("{:?}", cfg);
+        // TODO(cory): Ideally we should use serde to deserialize the input arguments
         let args: SQLFunctionArguments = planner.parse_function_args(
             inputs,
             &[
@@ -140,9 +137,6 @@ impl SQLFunction for HTTPConfigFunction {
         inputs: &[sqlparser::ast::FunctionArg],
         planner: &crate::planner::SQLPlanner,
     ) -> crate::error::SQLPlannerResult<daft_dsl::ExprRef> {
-        let deserializer = FunctionArgDeserializer::new(inputs, planner);
-        let cfg = deserializer.deserialize::<HTTPConfig>();
-        println!("cfg = {:?}", cfg);
         let args: SQLFunctionArguments =
             planner.parse_function_args(inputs, &["user_agent", "bearer_token"], 0)?;
 
@@ -161,608 +155,97 @@ impl SQLFunction for HTTPConfigFunction {
     }
 }
 
-struct FunctionArgDeserializer<'de> {
-    inputs: &'de [sqlparser::ast::FunctionArg],
-    planner: &'de crate::planner::SQLPlanner,
-    index: usize,
-}
+pub(crate) fn expr_to_iocfg(expr: &ExprRef) -> SQLPlannerResult<IOConfig> {
+    // TODO(CORY): use serde to deserialize this
+    let Expr::Literal(LiteralValue::Struct(entries)) = expr.as_ref() else {
+        unsupported_sql_err!("Invalid IOConfig");
+    };
 
-impl<'de> FunctionArgDeserializer<'de> {
-    fn new(
-        inputs: &'de [sqlparser::ast::FunctionArg],
-        planner: &'de crate::planner::SQLPlanner,
-    ) -> Self {
-        Self {
-            inputs,
-            planner,
-            index: 0,
+    macro_rules! get_value {
+        ($field:literal, $type:ident) => {
+            entries
+                .get(&Field::new($field, DataType::$type))
+                .and_then(|s| match s {
+                    LiteralValue::$type(s) => Some(Ok(s.clone())),
+                    LiteralValue::Null => None,
+                    _ => Some(Err(PlannerError::invalid_argument($field, "IOConfig"))),
+                })
+                .transpose()
+        };
+    }
+
+    let variant = get_value!("variant", Utf8)?
+        .expect("variant is required for IOConfig, this indicates a programming error");
+
+    match variant.as_ref() {
+        "s3" => {
+            let region_name = get_value!("region_name", Utf8)?;
+            let endpoint_url = get_value!("endpoint_url", Utf8)?;
+            let key_id = get_value!("key_id", Utf8)?;
+            let session_token = get_value!("session_token", Utf8)?.map(|s| s.into());
+            let access_key = get_value!("access_key", Utf8)?.map(|s| s.into());
+            let buffer_time = get_value!("buffer_time", UInt64)?;
+            let max_connections_per_io_thread =
+                get_value!("max_connections_per_io_thread", UInt32)?;
+            let retry_initial_backoff_ms = get_value!("retry_initial_backoff_ms", UInt64)?;
+            let connect_timeout_ms = get_value!("connect_timeout_ms", UInt64)?;
+            let read_timeout_ms = get_value!("read_timeout_ms", UInt64)?;
+            let num_tries = get_value!("num_tries", UInt32)?;
+            let retry_mode = get_value!("retry_mode", Utf8)?;
+            let anonymous = get_value!("anonymous", Boolean)?;
+            let use_ssl = get_value!("use_ssl", Boolean)?;
+            let verify_ssl = get_value!("verify_ssl", Boolean)?;
+            let check_hostname_ssl = get_value!("check_hostname_ssl", Boolean)?;
+            let requester_pays = get_value!("requester_pays", Boolean)?;
+            let force_virtual_addressing = get_value!("force_virtual_addressing", Boolean)?;
+            let profile_name = get_value!("profile_name", Utf8)?;
+            let default = S3Config::default();
+            let s3_config = S3Config {
+                region_name,
+                endpoint_url,
+                key_id,
+                session_token,
+                access_key,
+                credentials_provider: None,
+                buffer_time,
+                max_connections_per_io_thread: max_connections_per_io_thread
+                    .unwrap_or(default.max_connections_per_io_thread),
+                retry_initial_backoff_ms: retry_initial_backoff_ms
+                    .unwrap_or(default.retry_initial_backoff_ms),
+                connect_timeout_ms: connect_timeout_ms.unwrap_or(default.connect_timeout_ms),
+                read_timeout_ms: read_timeout_ms.unwrap_or(default.read_timeout_ms),
+                num_tries: num_tries.unwrap_or(default.num_tries),
+                retry_mode,
+                anonymous: anonymous.unwrap_or(default.anonymous),
+                use_ssl: use_ssl.unwrap_or(default.use_ssl),
+                verify_ssl: verify_ssl.unwrap_or(default.verify_ssl),
+                check_hostname_ssl: check_hostname_ssl.unwrap_or(default.check_hostname_ssl),
+                requester_pays: requester_pays.unwrap_or(default.requester_pays),
+                force_virtual_addressing: force_virtual_addressing
+                    .unwrap_or(default.force_virtual_addressing),
+                profile_name,
+            };
+
+            Ok(IOConfig {
+                s3: s3_config,
+                ..Default::default()
+            })
         }
-    }
+        "http" => {
+            let default = HTTPConfig::default();
+            let user_agent = get_value!("user_agent", Utf8)?.unwrap_or(default.user_agent);
+            let bearer_token = get_value!("bearer_token", Utf8)?.map(|s| s.into());
 
-    fn deserialize<T: Deserialize<'de>>(self) -> Result<T, DeserializerError> {
-        T::deserialize(self)
-    }
-
-    fn current_input(&self) -> Result<&sqlparser::ast::FunctionArg, DeserializerError> {
-        if self.index < self.inputs.len() {
-            Ok(&self.inputs[self.index])
-        } else {
-            Err(DeserializerError)
+            Ok(IOConfig {
+                http: HTTPConfig {
+                    user_agent,
+                    bearer_token,
+                },
+                ..Default::default()
+            })
         }
-    }
-    fn next_input(&mut self) -> Result<&sqlparser::ast::FunctionArg, DeserializerError> {
-        if self.index < self.inputs.len() {
-            let input = &self.inputs[self.index];
-            self.index += 1;
-            Ok(input)
-        } else {
-            Err(DeserializerError)
+        _ => {
+            unsupported_sql_err!("Unsupported IOConfig variant: {}", variant);
         }
-    }
-}
-
-#[derive(Debug)]
-struct DeserializerError;
-
-impl std::fmt::Display for DeserializerError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "deserializer error")
-    }
-}
-impl std::error::Error for DeserializerError {}
-impl serde::de::Error for DeserializerError {
-    fn custom<T: std::fmt::Display>(_msg: T) -> Self {
-        DeserializerError
-    }
-}
-impl<'de> serde::de::Deserializer<'de> for FunctionArgDeserializer<'de> {
-    type Error = DeserializerError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_bool<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_i8<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_i16<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_i32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_i64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_u8<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_u16<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_u32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_u64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_f32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_f64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_char<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_str<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_string<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_bytes<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_byte_buf<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_option<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_unit<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_unit_struct<V>(
-        self,
-        name: &'static str,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_newtype_struct<V>(
-        self,
-        name: &'static str,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_seq<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_tuple<V>(self, len: usize, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_tuple_struct<V>(
-        self,
-        name: &'static str,
-        len: usize,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_map<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_struct<V>(
-        self,
-        name: &'static str,
-        fields: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        let mut positional_args = HashMap::new();
-        let mut named_args = HashMap::new();
-        for (idx, arg) in self.inputs.iter().enumerate() {
-            match arg {
-                FunctionArg::Named {
-                    name,
-                    arg,
-                    operator: FunctionArgOperator::Assignment,
-                } => {
-                    if !fields.contains(&name.value.as_str()) {
-                        todo!()
-                    }
-                    named_args.insert(
-                        name.to_string(),
-                        self.planner.try_unwrap_function_arg_expr(&arg).unwrap(),
-                    );
-                }
-                FunctionArg::Unnamed(arg) => {
-                    todo!()
-                }
-                _ => todo!(),
-            }
-        }
-        // for field in fields {
-        //     if !named_args.contains_key(*field) {
-        //         named_args.insert(field.to_string(), null_lit());
-        //     }
-        // }
-
-        let res = visitor.visit_map(StructDeserializer {
-            fields,
-            named_args,
-            positional_args,
-            field_index: 0,
-        });
-        res.map_err(|e| e.into())
-    }
-
-    fn deserialize_enum<V>(
-        self,
-        name: &'static str,
-        variants: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_identifier<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_ignored_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: serde::de::Visitor<'de>,
-    {
-        todo!()
-    }
-}
-
-struct StructDeserializer {
-    fields: &'static [&'static str],
-    named_args: HashMap<String, ExprRef>,
-    positional_args: HashMap<usize, ExprRef>,
-    field_index: usize,
-}
-
-impl<'de, 'a> MapAccess<'de> for StructDeserializer {
-    type Error = DeserializerError;
-
-    fn next_key_seed<K>(&mut self, seed: K) -> Result<Option<K::Value>, Self::Error>
-    where
-        K: DeserializeSeed<'de>,
-    {
-        if self.field_index >= self.fields.len() {
-            Ok(None)
-        } else {
-            let field = self.fields[self.field_index];
-            seed.deserialize(field.into_deserializer()).map(Some)
-        }
-    }
-
-    fn next_value_seed<V>(&mut self, seed: V) -> Result<V::Value, Self::Error>
-    where
-        V: DeserializeSeed<'de>,
-    {
-        let field = self.fields[self.field_index];
-        self.field_index += 1;
-
-        if let Some(value) = self.named_args.get(field) {
-            let value = value.as_literal().unwrap().clone();
-
-            seed.deserialize(LiteralValueDeserializer { value: Some(value) })
-
-            // Handle named argument
-        } else if let Some(value) = self.positional_args.get(&(self.field_index - 1)) {
-            // Handle positional argument
-            todo!()
-        } else {
-            seed.deserialize(LiteralValueDeserializer { value: None })
-        }
-    }
-}
-
-struct LiteralValueDeserializer {
-    value: Option<LiteralValue>,
-}
-
-impl<'de> Deserializer<'de> for LiteralValueDeserializer {
-    type Error = DeserializerError;
-
-    fn deserialize_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_bool<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_bool(self.value.unwrap().as_bool().unwrap())
-    }
-
-    fn deserialize_i8<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_i8(self.value.unwrap().as_i32().unwrap() as i8)
-    }
-
-    fn deserialize_i16<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_i16(self.value.unwrap().as_i32().unwrap() as i16)
-    }
-
-    fn deserialize_i32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_i32(self.value.unwrap().as_i32().unwrap())
-    }
-
-    fn deserialize_i64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_i64(self.value.unwrap().as_i64().unwrap())
-    }
-
-    fn deserialize_u8<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_u8(self.value.unwrap().as_u32().unwrap() as u8)
-    }
-
-    fn deserialize_u16<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_u16(self.value.unwrap().as_u32().unwrap() as u16)
-    }
-
-    fn deserialize_u32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_u32(self.value.unwrap().as_u32().unwrap())
-    }
-
-    fn deserialize_u64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_u64(self.value.unwrap().as_u64().unwrap())
-    }
-
-    fn deserialize_f32<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_f32(self.value.unwrap().as_f64().unwrap() as f32)
-    }
-
-    fn deserialize_f64<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_f64(self.value.unwrap().as_f64().unwrap())
-    }
-
-    fn deserialize_char<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_str<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_str(self.value.unwrap().as_str().unwrap_or(""))
-    }
-
-    fn deserialize_string<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        match self.value {
-            None => visitor.visit_str(""),
-            Some(LiteralValue::Null) => visitor.visit_none(),
-            Some(LiteralValue::Utf8(s)) => visitor.visit_string(s),
-            other => panic!("Expected string, got {:?}", other),
-        }
-    }
-
-    fn deserialize_bytes<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_bytes(self.value.unwrap().as_binary().unwrap())
-    }
-
-    fn deserialize_byte_buf<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_byte_buf(self.value.unwrap().as_binary().unwrap().to_vec())
-    }
-
-    fn deserialize_option<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        match self.value {
-            None => visitor.visit_unit(),
-            Some(LiteralValue::Null) => visitor.visit_none(),
-            _ => visitor.visit_some(self),
-        }
-    }
-
-    fn deserialize_unit<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        visitor.visit_unit()
-    }
-
-    fn deserialize_unit_struct<V>(
-        self,
-        name: &'static str,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_newtype_struct<V>(
-        self,
-        name: &'static str,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_seq<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_tuple<V>(self, len: usize, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_tuple_struct<V>(
-        self,
-        name: &'static str,
-        len: usize,
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_map<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_struct<V>(
-        self,
-        name: &'static str,
-        fields: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_enum<V>(
-        self,
-        name: &'static str,
-        variants: &'static [&'static str],
-        visitor: V,
-    ) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_identifier<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
-    }
-
-    fn deserialize_ignored_any<V>(self, visitor: V) -> Result<V::Value, Self::Error>
-    where
-        V: Visitor<'de>,
-    {
-        todo!()
     }
 }
