+++ src/daft-sql/src/table_provider/read_parquet.rs
@@ -1,0 +1,77 @@
+use daft_core::prelude::TimeUnit;
+use daft_plan::{LogicalPlanBuilder, ParquetScanBuilder};
+use sqlparser::ast::TableFunctionArgs;
+
+use super::SQLTableFunction;
+use crate::{
+    error::{PlannerError, SQLPlannerResult},
+    functions::SQLFunctionArguments,
+    modules::config::expr_to_iocfg,
+    planner::SQLPlanner,
+};
+
+pub(super) struct ReadParquetFunction;
+
+impl TryFrom<SQLFunctionArguments> for ParquetScanBuilder {
+    type Error = PlannerError;
+
+    fn try_from(args: SQLFunctionArguments) -> Result<Self, Self::Error> {
+        let glob_paths: String = args.try_get_positional(0)?.ok_or_else(|| {
+            PlannerError::invalid_operation("path is required for `read_parquet`")
+        })?;
+        let infer_schema = args.try_get_named("infer_schema")?.unwrap_or(true);
+        let coerce_int96_timestamp_unit =
+            args.try_get_named::<String>("coerce_int96_timestamp_unit")?;
+        let coerce_int96_timestamp_unit: TimeUnit = coerce_int96_timestamp_unit
+            .as_deref()
+            .unwrap_or("nanoseconds")
+            .parse()
+            .map_err(|_| {
+                PlannerError::invalid_argument("coerce_int96_timestamp_unit", "read_parquet")
+            })?;
+        let chunk_size = args.try_get_named("chunk_size")?;
+        let multithreaded = args.try_get_named("multithreaded")?.unwrap_or(true);
+
+        let field_id_mapping = None; // TODO
+        let row_groups = None; // TODO
+        let schema = None; // TODO
+        let io_config = args.get_named("io_config").map(expr_to_iocfg).transpose()?;
+
+        Ok(Self {
+            glob_paths: vec![glob_paths],
+            infer_schema,
+            coerce_int96_timestamp_unit,
+            field_id_mapping,
+            row_groups,
+            chunk_size,
+            io_config,
+            multithreaded,
+            schema,
+        })
+    }
+}
+
+impl SQLTableFunction for ReadParquetFunction {
+    fn plan(
+        &self,
+        planner: &SQLPlanner,
+        args: &TableFunctionArgs,
+    ) -> SQLPlannerResult<LogicalPlanBuilder> {
+        let builder: ParquetScanBuilder = planner.plan_function_args(
+            args.args.as_slice(),
+            &[
+                "infer_schema",
+                "coerce_int96_timestamp_unit",
+                "chunk_size",
+                "multithreaded",
+                // "schema",
+                // "field_id_mapping",
+                // "row_groups",
+                "io_config",
+            ],
+            1, // 1 positional argument (path)
+        )?;
+
+        builder.finish().map_err(From::from)
+    }
+}
